backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.9
openai==1.61.0
httpx==0.27.2

# File handling
pypdf==4.3.1
pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
